# users/openapi.yaml
openapi: 3.0.0
info:
  title: User Service API
  version: 2.0.0
  description: API for managing users, authentication and games

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    Coordinates:
      type: object
      required: [x, y, z]
      properties:
        x:
          type: integer
          example: 1
        y:
          type: integer
          example: 1
        z:
          type: integer
          example: 1

    Move:
      type: object
      properties:
        move_number:
          type: integer
          example: 1
        player:
          type: string
          enum: [B, R]
          example: B
        coordinates:
          $ref: '#/components/schemas/Coordinates'
        yen_state:
          type: string
          nullable: true
          example: "B/.B/RB./B..R"
        created_at:
          type: string
          format: date-time

    VsPlayerStats:
      type: object
      properties:
        wins:
          type: integer
          example: 2
        losses:
          type: integer
          example: 1
        draws:
          type: integer
          example: 0

    BotStat:
      type: object
      description: Stats for a single bot strategy, normalized with name and difficulty
      properties:
        name:
          type: string
          example: "random"
        difficulty:
          type: string
          example: "Easy 😄"
        wins:
          type: integer
          example: 3
        losses:
          type: integer
          example: 1
        draws:
          type: integer
          example: 0

    Statistics:
      type: object
      properties:
        total_games:
          type: integer
          example: 10
        total_wins:
          type: integer
          example: 6
        total_losses:
          type: integer
          example: 4
        total_draws:
          type: integer
          example: 1
        vs_player:
          $ref: '#/components/schemas/VsPlayerStats'
        vs_bots:
          type: array
          items:
            $ref: '#/components/schemas/BotStat'
          example:
            - { name: "random",    difficulty: "Easy 😄",   wins: 1, losses: 0, draws: 0 }
            - { name: "defensive", difficulty: "Medium 😐", wins: 2, losses: 1, draws: 0 }
            - { name: "ncts",      difficulty: "Hard 😈",   wins: 0, losses: 3, draws: 1 }

    StrategyOption:
      type: object
      properties:
        name:
          type: string
          example: "Random"
        difficulty:
          type: string
          example: "Easy 😄"

    VariantOption:
      type: object
      properties:
        name:
          type: string
          example: "Classic Y"

    GameOptions:
      type: object
      properties:
        strategies:
          type: array
          items:
            $ref: '#/components/schemas/StrategyOption'
          example:
            - { name: "Random",    difficulty: "Easy 😄"   }
            - { name: "Defensive", difficulty: "Medium 😐" }
            - { name: "NCTS",      difficulty: "Hard 😈"   }
        variants:
          type: array
          items:
            $ref: '#/components/schemas/VariantOption'
          example:
            - { name: 'Classic Y' }
            - { name: 'Master Y (coming soon)' }
            - { name: 'Pie Rule (coming soon)' }

    LeaderboardEntry:
      type: object
      properties:
        username:
          type: string
          example: "nacho"
        total_wins:
          type: integer
          example: 42
        total_games:
          type: integer
          example: 50

    BotLeaderboardEntry:
      type: object
      properties:
        username:
          type: string
          example: "nacho"
        wins:
          type: integer
          example: 15

    Leaderboard:
      type: object
      properties:
        overall:
          type: array
          description: Top 10 players by total wins
          items:
            $ref: '#/components/schemas/LeaderboardEntry'
        vs_bots:
          type: object
          description: Top 10 players per bot strategy
          properties:
            random:
              type: array
              items:
                $ref: '#/components/schemas/BotLeaderboardEntry'
            defensive:
              type: array
              items:
                $ref: '#/components/schemas/BotLeaderboardEntry'
            ncts:
              type: array
              items:
                $ref: '#/components/schemas/BotLeaderboardEntry'

    Game:
      type: object
      properties:
        _id:
          type: string
          example: "64b1f2e3c4d5e6f7a8b9c0d1"
        player_id:
          type: string
          example: "64b1f2e3c4d5e6f7a8b9c0d2"
        game_type:
          type: string
          enum: [BOT, PLAYER]
          example: BOT
        name_of_enemy:
          type: string
          nullable: true
          example: null
        board_size:
          type: integer
          example: 7
        strategy:
          type: string
          example: "random"
        current_turn:
          type: string
          enum: [B, R]
          example: B
        status:
          type: string
          enum: [IN_PROGRESS, FINISHED]
          example: IN_PROGRESS
        result:
          type: string
          enum: [WIN, LOSS, DRAW]
          nullable: true
          example: null
        duration_seconds:
          type: integer
          example: 0
        yen_final_state:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        moves:
          type: array
          items:
            $ref: '#/components/schemas/Move'

    User:
      type: object
      properties:
        _id:
          type: string
          example: "64b1f2e3c4d5e6f7a8b9c0d2"
        username:
          type: string
          example: "nacho"
        created_at:
          type: string
          format: date-time
        statistics:
          $ref: '#/components/schemas/Statistics'

paths:

  # Auth

  /createuser:
    post:
      summary: Register a new user
      description: Creates a new user with a hashed password.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                  example: "nacho"
                password:
                  type: string
                  example: "securepassword"
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Welcome nacho!"
                  userId:
                    type: string
                    example: "64b1f2e3c4d5e6f7a8b9c0d2"
        '400':
          description: Username and password are required, or input is not a string
        '409':
          description: Username already taken
        '500':
          description: Internal server error

  /login:
    post:
      summary: Login
      description: Authenticates a user and returns a JWT token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                  example: "nacho"
                password:
                  type: string
                  example: "securepassword"
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                  username:
                    type: string
                    example: "nacho"
                  userId:
                    type: string
                    example: "64b1f2e3c4d5e6f7a8b9c0d2"
        '400':
          description: Username and password are required, or input is not a string
        '401':
          description: Invalid credentials
        '500':
          description: Internal server error

  /exists/{username}:
    get:
      summary: Check if username is taken
      description: Returns whether a username is already registered. Public endpoint, no authentication required.
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
          example: "nacho"
      responses:
        '200':
          description: Username availability
          content:
            application/json:
              schema:
                type: object
                properties:
                  exists:
                    type: boolean
                    example: false
        '500':
          description: Internal server error

  # Users

  /leaderboard:
    get:
      summary: Get leaderboard
      description: Returns top 10 players overall and per bot strategy. Public endpoint, no authentication required.
      responses:
        '200':
          description: Leaderboard with overall ranking and per-bot rankings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Leaderboard'
        '500':
          description: Internal server error

  /users/{id}:
    get:
      summary: Get user profile
      description: Returns user profile and statistics. Requires authentication.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: "64b1f2e3c4d5e6f7a8b9c0d2"
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: No token provided or invalid token
        '404':
          description: User not found
        '500':
          description: Internal server error

  /users/{id}/history:
    get:
      summary: Get user game history
      description: Returns all games played by the user ordered by date (without moves array). Requires authentication.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: "64b1f2e3c4d5e6f7a8b9c0d2"
      responses:
        '200':
          description: List of games without moves array
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Game'
        '401':
          description: No token provided or invalid token
        '500':
          description: Internal server error

  # Games

  /games/options:
    get:
      summary: Get available game options
      description: Returns the available strategies and game variants. Public endpoint, no authentication required.
      responses:
        '200':
          description: Available game options
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameOptions'
        '500':
          description: Internal server error

  /play:
    post:
      summary: Public bot play — no auth or game id required
      description: >
        Open source endpoint for external users. Send a board position and optional bot_id
        to get the bot's next move. No authentication or internal game id needed.
        Defaults to the hardest bot (ncts) if bot_id is not provided.
        board_size defaults to 5 if not provided.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [position]
              properties:
                position:
                  type: string
                  nullable: true
                  description: Current board in YEN notation. Send null for the first move.
                  example: "B/.B/RB./B..R"
                bot_id:
                  type: string
                  enum: [random, defensive, ncts]
                  default: ncts
                  description: "Bot strategy to use. Defaults to ncts (hardest) if not provided."
                  example: "ncts"
                board_size:
                  type: integer
                  default: 5
                  example: 5
      responses:
        '200':
          description: Bot move returned from Gamey
          content:
            application/json:
              schema:
                type: object
                properties:
                  coordinates:
                    $ref: '#/components/schemas/Coordinates'
                  yen_state:
                    type: string
                    example: "R/.B/RB./B..R"
                  winner:
                    type: string
                    enum: [B, R]
                    nullable: true
                    example: null
        '400':
          description: position is required
        '502':
          description: Gamey service returned an error
        '503':
          description: Gamey service unreachable

  /games:
    post:
      summary: Create a new game
      description: Creates a new game for the authenticated user. Starting turn (B or R) is assigned randomly. Requires authentication.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [board_size]
              properties:
                board_size:
                  type: integer
                  example: 7
                game_type:
                  type: string
                  enum: [BOT, PLAYER]
                  default: BOT
                  example: BOT
                name_of_enemy:
                  type: string
                  description: Required when game_type is PLAYER
                  example: "Tobias"
                strategy:
                  type: string
                  enum: [random, defensive, ncts]
                  default: random
                  description: "Difficulty is derived automatically: random=Easy 😄, defensive=Medium 😐, ncts=Hard 😈"
                  example: "random"
      responses:
        '201':
          description: Game created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Game'
        '400':
          description: board_size is required, or name_of_enemy is required for PLAYER games
        '401':
          description: No token provided or invalid token
        '500':
          description: Internal server error

  /games/{id}:
    get:
      summary: Get game state
      description: Returns the full game state including all moves. Requires authentication.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: "64b1f2e3c4d5e6f7a8b9c0d1"
      responses:
        '200':
          description: Game state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Game'
        '401':
          description: No token provided or invalid token
        '404':
          description: Game not found
        '500':
          description: Internal server error

  /games/{id}/move:
    post:
      summary: Submit a player move
      description: >
        Saves a player move. Only coordinates are required — the player (B or R) is inferred from
        current_turn. The backend calls Gamey to compute the new yen_state and check for a winner.
        If Gamey returns a winner the game is automatically finished and statistics updated.
        Returns the full updated game state (check status and result fields). Requires authentication.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: "64b1f2e3c4d5e6f7a8b9c0d1"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [coordinates]
              properties:
                coordinates:
                  $ref: '#/components/schemas/Coordinates'
      responses:
        '201':
          description: Move saved — returns full updated game state. If game is over, status=FINISHED and result=WIN or LOSS.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Game'
        '400':
          description: coordinates (x, y, z) are required, or game is already finished
        '401':
          description: No token provided or invalid token
        '404':
          description: Game not found
        '502':
          description: Gamey compute service returned an error
        '503':
          description: Gamey compute service unreachable
        '500':
          description: Internal server error

  /games/{id}/play:
    get:
      summary: Request bot move
      description: >
        Reads the yen_state from the last saved move in DB, and forwards it along with strategy
        and board_size to Gamey. Gamey returns the bot's chosen coordinates, yen_state, and winner.
        The bot move is saved automatically. If Gamey returns a winner the game is automatically
        finished and statistics updated. Returns the full updated game state. Requires authentication.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: "64b1f2e3c4d5e6f7a8b9c0d1"
      responses:
        '201':
          description: Bot move saved — returns full updated game state. If game is over, status=FINISHED and result=WIN or LOSS.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Game'
        '400':
          description: Game is already finished
        '401':
          description: No token provided or invalid token
        '404':
          description: Game not found
        '502':
          description: Gamey service returned an error
        '503':
          description: Gamey service unreachable

  /games/{id}/undo:
    post:
      summary: Undo last move
      description: Removes the last move and reverts the turn. Only allowed in PLAYER vs PLAYER games that are still IN_PROGRESS. Requires authentication.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: "64b1f2e3c4d5e6f7a8b9c0d1"
      responses:
        '200':
          description: Last move removed — returns updated game state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Game'
        '400':
          description: No moves to undo, game is already finished, or game is not a PLAYER vs PLAYER game
        '401':
          description: No token provided or invalid token
        '404':
          description: Game not found
        '500':
          description: Internal server error

  /games/{id}/moves:
    get:
      summary: Get all moves of a game
      description: Returns all moves ordered by move_number for replay purposes. Requires authentication.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: "64b1f2e3c4d5e6f7a8b9c0d1"
      responses:
        '200':
          description: List of moves ordered by move_number
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Move'
        '401':
          description: No token provided or invalid token
        '404':
          description: Game not found
        '500':
          description: Internal server error

  /games/{id}/finish:
    put:
      summary: Finish a game
      description: Marks a game as FINISHED manually — used when a user quits (DRAW). WIN/LOSS are handled automatically by the backend when Gamey returns a winner. Requires authentication.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: "64b1f2e3c4d5e6f7a8b9c0d1"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [result]
              properties:
                result:
                  type: string
                  enum: [WIN, LOSS, DRAW]
                  example: "DRAW"
                yen_final_state:
                  type: string
                  nullable: true
                  example: "B/.B/RB./B..R"
                duration_seconds:
                  type: integer
                  example: 120
      responses:
        '200':
          description: Game finished
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Game'
        '400':
          description: result is required, or game is already finished
        '401':
          description: No token provided or invalid token
        '404':
          description: Game not found
        '500':
          description: Internal server error